Added create user API

+++ internal/types/types.go
@@ -1,8 +1,9 @@
 package types
 
 type Student struct {
-	Id    int
+	Id    int64
 	Name  string `validate:"required"`
 	Email string `validate:"required,email"`
 	Age   int
 }
+
